rename directory to switchChainErrorModal from switchChainModal
